Created a resposive design using @media query
