+++ client/src/components/style.css
@@ -388,3 +388,77 @@ input[type=password]:focus {
 }
 
 /* Login Component Style End */
+
+
+/* Responsive Design Style Start */
+
+@media (max-width: 768px){
+  .navbar-container{
+    flex-direction: column-reverse;
+  }
+  nav{
+    width: 100%;
+    height: 10vh;
+    flex-direction: row;
+    border-top: 1px solid rgb(81, 81, 81);
+    position: sticky;
+    bottom: 0;
+  }
+  .navbar-section-menus-area{
+    flex-direction: row;
+  }
+  .content{
+    width: 100%;
+    height: 190vh;
+    z-index: -1;
+  }
+  .onboarding-component-onboarding-screens-content-area{
+    width: 100%;
+  }
+  .onboarding-component-onboarding-screens-content-area{
+    flex-direction: column;
+  }
+  .onboarding-component-onboarding-screens{
+    width: 90vw;
+    height: 100%;
+  }
+  .onboarding-component-slider-slide img{
+    width: 90vw;
+    height: 40vh;
+  }
+  .onboarding-component-slides-text-area{
+    width: 90vw;
+  }
+  .onboarding-component-demo-music-player{
+    width: 90vw;
+  }
+  .onboarding-component-ads-content-area{
+    width: 90vw;
+  }
+  .onboarding-component-ads-content-area img{
+    width: 90vw;
+  }
+  .login-component{
+    flex-direction: column;
+    width: 80vw;
+    height: 90vh;
+  }
+  .login-component-left-image-area{
+      width: 100%;
+      height: 20vh;
+  }
+  .login-component-right-form-area{
+    width: 80vw;
+  }
+  .login-component-username-area input{
+    width: 65vw;
+  }
+  .login-component-form-password-inputs input{
+    width: 65vw;
+  }
+  .login-component-random-quote{
+    width: 55vw;
+  }
+}
+
+/* Responsive Design Style End */
